+++ examples/pdl1-rfd/run-local-simple.sh
@@ -4,17 +4,12 @@ PIPELINE_DIR=../../
 
 DATESTAMP=$(date +%Y%m%d_%H%M%S)
 
-nextflow run ${PIPELINE_DIR}/partial.nf  \
+nextflow run ${PIPELINE_DIR}/main.nf  \
   --input_pdb 'input/*.pdb' \
   --outdir results \
   --contigs "[A18-132/0 65-120]" \
   --hotspot_res "A56" \
-  --rfd_partial_per_binder=1 \
-  --rfd_batch_size=1 \
-  --rfd_partial_T "5,15" \
-  --pmpnn_seqs_per_struct=1 \
-  --pmpnn_relax_cycles=1 \
-  --rfd_filters="rg<=20" \
+  --rfd_n_designs=4 \
   -profile local \
   -resume \
   -with-report results/logs/report_${DATESTAMP}.html \